refactor: clean up unused imports and improve component structure

+++ src/context/AuthContext.tsx
@@ -1,7 +1,6 @@
 /* eslint-disable react-refresh/only-export-components */
 import {
     sendPasswordResetEmail,
-    signInWithEmailAndPassword,
     signOut,
     signInWithPopup,
     GoogleAuthProvider,
@@ -13,33 +12,39 @@ import {
     linkWithCredential,
     signInAnonymously,
     onAuthStateChanged,
-    User
+    User,
+    signInWithEmailAndPassword,
+    getIdToken,
 } from "firebase/auth";
 import { createContext, useCallback, useEffect, useState } from "react";
 import { ReactNode } from "react";
 import { auth } from "@/utils/firebase";
 import { doc, setDoc, updateDoc } from "firebase/firestore";
 import { db } from "@/utils/firebase";
-import { AbsoluteCenter, Spinner } from "@chakra-ui/react";
+import { AbsoluteCenter, Spinner, Text } from "@chakra-ui/react";
 
 interface AuthContextProps {
+    getUserToken: () => Promise<string>;
+    setCurrentUser:
+    React.Dispatch<React.SetStateAction<User | null>>;
+    isAuthenticated: boolean;
+    subscribed: boolean;
+    isEmailNotificationEnabled: boolean;
     currentUser: User | null;
-    loginWithEmailAndPassword: (email: string, password: string) => Promise<void>;
     loginWithGoogle: () => Promise<void>;
     signUpWithEmailAndPassword: (
         email: string,
         password: string
     ) => Promise<void>;
     signOut: () => Promise<void>;
-    getUserToken: () => Promise<string>;
     handleRecoverPassword: (email: string) => Promise<void>;
-    updatePhotoURLInContext: (newPhotoURL: string) => void;
     sendPasswordReset: (email: string) => Promise<void>;
     handleEmailChange: (currentPassword: string, email: string) => Promise<void>;
     updateUserName: (firstName: string, secondName: string) => Promise<void>;
     updateEmailNotificationPreference: (preference: boolean) => Promise<void>;
     deleteUserAccount: (currentPassword: string) => Promise<void>;
     upgradeFromAnonymous: (email: string, password: string) => Promise<User>;
+    simpleLogin: (email: string, password: string) => Promise<User>;
 };
 
 
@@ -48,7 +53,9 @@ export const AuthContext = createContext<AuthContextProps | null>(null);
 export function AuthProvider({ children }: { children: ReactNode }) {
     const [currentUser, setCurrentUser] = useState<User | null>(null);
     const [loading, setLoading] = useState(true);
-
+    const [isAuthenticated, setIsAuthenticated] = useState(false);
+    const [isEmailNotificationEnabled,] = useState(true);
+    const [subscribed,] = useState(true);
     useEffect(() => {
         const unsubscribe = onAuthStateChanged(auth, async (user) => {
             if (user) {
@@ -71,21 +78,14 @@ export function AuthProvider({ children }: { children: ReactNode }) {
     }, []);
 
     async function upgradeFromAnonymous(email: string, password: string) {
-        // user atual (possivelmente anônimo)
         const user = auth.currentUser;
-
         if (!user) {
             throw new Error("Nenhum usuário autenticado para fazer upgrade.");
         }
-
-        // Cria credenciais
         const credential = EmailAuthProvider.credential(email, password);
-
-        // Faz o link
         const result = await linkWithCredential(user, credential);
-        // Agora o user não é mais anônimo
         setCurrentUser(result.user);
-
+        setIsAuthenticated(true);
         return result.user;
     }
 
@@ -107,8 +107,8 @@ export function AuthProvider({ children }: { children: ReactNode }) {
     const deleteUserAccount = async (currentPassword: string) => {
         console.log("Deleting user account...");
         try {
-            if (currentUser) {
-                const credential = EmailAuthProvider.credential(user.email, currentPassword);
+            if (currentUser && currentUser.email) {
+                const credential = EmailAuthProvider.credential(currentUser.email, currentPassword);
                 await reauthenticateWithCredential(currentUser, credential);
                 await deleteUser(currentUser);
             }
@@ -118,17 +118,20 @@ export function AuthProvider({ children }: { children: ReactNode }) {
     }
 
     const handleEmailChange = async (currentPassword: string, newEmail: string) => {
-        if (currentUser) {
-            try {
-                const credential = EmailAuthProvider.credential(user.email, currentPassword);
-                await reauthenticateWithCredential(currentUser, credential);
-                await verifyBeforeUpdateEmail(currentUser, newEmail);
-                await updateEmail(currentUser, newEmail);
-            } catch (error) {
-                console.error("Erro ao atualizar email:", error);
-            }
+        if (!currentUser || !currentUser.email) return;
+        try {
+            const credential = EmailAuthProvider.credential(
+                currentUser.email,
+                currentPassword
+            );
+            await reauthenticateWithCredential(currentUser, credential);
+            await verifyBeforeUpdateEmail(currentUser, newEmail);
+            await updateEmail(currentUser, newEmail);
+        } catch (error) {
+            console.error("Erro ao atualizar email:", error);
         }
     };
+
 
     const updateUserName = async (firstName: string, secondName: string) => {
         if (currentUser) {
@@ -142,28 +145,71 @@ export function AuthProvider({ children }: { children: ReactNode }) {
     };
 
 
-    const loginWithEmailAndPassword = useCallback(async (email: string, password: string) => {
-        await signInWithEmailAndPassword(auth, email, password);
-    }, []);
-
+    /**
+       * Login com Google (faz link automático se o user for anônimo,
+       * ou simplesmente "login" se não for).
+       */
     const loginWithGoogle = useCallback(async () => {
         const provider = new GoogleAuthProvider();
-        await signInWithPopup(auth, provider);
+        const user = auth.currentUser;
+        if (user?.isAnonymous) {
+            // Se é anônimo, linkamos
+            await signInWithPopup(auth, provider);
+            setIsAuthenticated(true);
+        } else {
+            // Se não é anônimo, também funciona, mas vira "multi-provider".
+            await signInWithPopup(auth, provider);
+        }
     }, []);
+
+    /**
+     * Função responsável por logar o usuário.
+     * @param email Identificador do Usuário
+     * @param password Senha de Login do Usuário
+     * @returns 
+     */
+    async function simpleLogin(email: string, password: string) {
+        try {
+            const result = await signInWithEmailAndPassword(auth, email, password);
+
+            // Configurar o usuário atual e o estado de autenticação
+            setCurrentUser(result.user);
+            setIsAuthenticated(true);
+
+            return result.user;
+        } catch (error) {
+            const authError = error as { code?: string; message: string };
+
+            // Tratar diferentes tipos de erros
+            if (authError.code === "auth/user-not-found") {
+                throw new Error("Usuário não encontrado. Verifique seu e-mail.");
+            } else if (authError.code === "auth/wrong-password") {
+                throw new Error("Senha incorreta. Tente novamente.");
+            } else {
+                console.error("Erro ao realizar login:", error);
+                throw new Error("Erro ao fazer login. Tente novamente.");
+            }
+        }
+    }
+
 
     const updateEmailNotificationPreference = async (preference: boolean) => {
         try {
-            const userRef = doc(db, "users", user.uid);
-            await updateDoc(userRef, { emailNotifications: preference });
+            if (currentUser) {
+                const userRef = doc(db, "users", currentUser.uid);
+                await updateDoc(userRef, { emailNotifications: preference });
+            }
             console.log("Preferência de notificação atualizada com sucesso.");
         } catch (error) {
             console.error("Erro ao atualizar preferência de notificação:", error);
         }
     };
 
+    // TODO : Create a log-out page
     const signOutFn = useCallback(async () => {
         await signOut(auth);
-        window.location.reload();
+        setIsAuthenticated(false)
+        window.location.assign("/");
     }, []);
 
     const handleRecoverPassword = useCallback(async (email: string) => {
@@ -178,9 +224,16 @@ export function AuthProvider({ children }: { children: ReactNode }) {
         }
     }, []);
 
+    const getUserToken = useCallback(async () => {
+        if (auth.currentUser) {
+            return await getIdToken(auth.currentUser);
+        }
+        return "";
+    }, []);
+
     const value = {
         currentUser,
-        loginWithEmailAndPassword,
+        isAuthenticated,
         loginWithGoogle,
         signUpWithEmailAndPassword,
         signOut: signOutFn,
@@ -190,12 +243,17 @@ export function AuthProvider({ children }: { children: ReactNode }) {
         updateUserName,
         updateEmailNotificationPreference,
         deleteUserAccount,
-        upgradeFromAnonymous
+        upgradeFromAnonymous,
+        simpleLogin,
+        isEmailNotificationEnabled,
+        subscribed,
+        setCurrentUser,
+        getUserToken
     };
 
     return (
         <AuthContext.Provider value={value}>
-            {!loading ? children : <AbsoluteCenter><Spinner>Loading...</Spinner></AbsoluteCenter>}
+            {!loading ? children : <AbsoluteCenter><Spinner size="xl" /><Text fontSize="5xl">Chargement...</Text></AbsoluteCenter>}
         </AuthContext.Provider>
     );
 }
